Demonstrate some fundamentals in exercise 3
This was me learning about a few basics:
* `grid-template-columns`
* `grid-template-rows`
* `grid-gap`

These seem like the basics I'd need to use, and probably enough to get
me a real long way with CSS grid anyway. Terms like "track" still seem
unnatural to me, but I'm hoping to get more comfortable with them as
time goes on.

I'm also looking forward to understanding when to use CSS grid vs when
to use flexbox. I've heard flexbox is only supposed to be used when
you're talking about aligning multiple items on a single axis, but I
want to hear something a little more nuanced and convincing.