Finish exercise 20
WHEW! That was a biggie. Lots of stuff I wasn't really comfortable with
yet, but I'm hoping that as I keep doing these exercises I get more
comfortable with grid. It seems like there's a bunch of techniques to
memorize, and right now I'm wondering the difference between
`justify-content` and `justify-items` -- I'm sure this was covered, but
I forget it. It seems to me like we should've been doing more baby
examples along the way to this thing.

Maybe down the road I should do this exercise again? I don't know the
best way to reinforce the most useful CSS grid techniques.