Complete exercise 22
Oh boy, a lot going on here too! I followed along with Wes as he did
everything (lots of styling unrelated to grid going on in this pup) and
had a rough idea as to how to do most of the stuff in CSS grid.

Some of my takeaways here were:
Work from the outside in
Expect some common requests and memorize the techniques

By working from the outside in, I mean systematically approach the
layout, broad strokes to fine strokes.

By expecting common requests, I mean there are probably just a dozen
things that designers will request regularly. Maybe it's three
side-by-side boxes, maybe it's a bunch of things that are auto sized and
the front takes up the rest, maybe it's some boxes that evenly split an
area... my point is that there's probably really just a handful of
requests, so rather than memorize all the possible techniques first,
maybe start by drilling the common stuff.