Complete exercise 23
This was quick and easy. I followed along with the video cause I didn't
really have any way of knowing what to do, but it seems nice to be able
to go Bootstrap-style if you want to, if for the very least you've got a
team that's familiar with that approach.