Complete exercise 25
WOOHOO! ALL DONE! This one was pretty easy and fun, I think what maybe
would make sense for the other exercises would be to have Wes prompt the
watcher to say "Okay, so now we want to [X], so how would you do that?"
since there's so much grid stuff mixed in with non-grid stuff... and
sometimes, I might not even know what it is I'm trying to do.

I think at a certain point I'll be able to break down the building into
discrete steps that I can understand, but I'm not there yet. But maybe I
can solve for a specific problem when it's framed as "Hey can you make
these three elements fill their parent up with the same width?" or
whatever